Extract common code for JUnit4 and JUnit5 CoroutinesTimeout
