+++ kotlinx-coroutines-debug/build.gradle
@@ -18,16 +18,12 @@ configurations {
     configureKotlinJvmPlatform(shadow)
 }
 
-test {
-    useJUnitPlatform()
-}
-
 dependencies {
     compileOnly "junit:junit:$junit_version"
     compileOnly "org.junit.jupiter:junit-jupiter-api:$junit5_version"
     testCompile "org.junit.jupiter:junit-jupiter-api:$junit5_version"
     testCompile "org.junit.jupiter:junit-jupiter-engine:$junit5_version"
-    testRuntimeOnly "org.junit.vintage:junit-vintage-engine:$junit5_version"
+    testCompile "org.junit.platform:junit-platform-testkit:1.7.0"
     shadowDeps "net.bytebuddy:byte-buddy:$byte_buddy_version"
     shadowDeps "net.bytebuddy:byte-buddy-agent:$byte_buddy_version"
     compileOnly "io.projectreactor.tools:blockhound:$blockhound_version"
@@ -44,6 +40,10 @@ if (rootProject.ext.jvm_ir_enabled) {
             excludeTestsMatching('kotlinx.coroutines.debug.DebugProbesTest')
         }
     }
+}
+
+java {
+    disableAutoTargetJvm()
 }
 
 jar {
