+++ kotlinx-coroutines-debug/test/junit5/CoroutinesTimeoutTest.kt
@@ -3,20 +3,43 @@
  */
 
 package junit5
-import kotlinx.coroutines.debug.junit5.*
-import org.junit.jupiter.api.*
 
-@CoroutinesTimeout(6)
+import org.assertj.core.api.*
+import org.junit.*
+import org.junit.platform.engine.discovery.DiscoverySelectors.*
+import org.junit.platform.testkit.engine.*
+import org.junit.platform.testkit.engine.EventConditions.*
+import org.junit.platform.testkit.engine.TestExecutionResultConditions.*
+import org.junit.runners.model.*
+
 class CoroutinesTimeoutTest {
 
-    @CoroutinesTimeout(5)
+    // note that this test is run using JUnit4 in order not to mix the testing systems.
     @Test
-    fun test() {
-
-    }
-
-    @Test
-    fun test2() {
-
+    fun testCoroutinesTimeout() {
+        EngineTestKit.engine("junit-jupiter")
+            .selectors(selectClass(CoroutinesTimeoutSimpleTest::class.java))
+            .execute()
+            .testEvents()
+            .assertThatEvents()
+            .testFinishedSuccessfully("ignoresClassTimeout")
+            .testFinishedSuccessfully("fitsInClassTimeout")
+            .testTimedOut("usesClassTimeout1", 100)
+            .testTimedOut("usesClassTimeout2", 100)
+            .testTimedOut("usesMethodTimeout", 200)
     }
 }
+
+private fun ListAssert<Event>.testFinishedSuccessfully(testName: String): ListAssert<Event> =
+    haveExactly(1, event(
+        test(testName),
+        finishedSuccessfully()
+    ))
+
+private fun ListAssert<Event>.testTimedOut(testName: String, after: Int): ListAssert<Event> =
+    haveExactly(1, event(
+        test(testName),
+        finishedWithFailure(instanceOf(TestTimedOutException::class.java), message {
+            it.contains(Regex("\\b$after\\b"))
+        })
+    ))
